requirements.txt
mcp>=1.0.0
googlesearch-python>=1.2.3

